refactor: 스크리닝 디테일 bottom button 리팩토링
