+++ src/screens/screening/detail/components/bottomDetailButton/index.ts
@@ -1,0 +1,2 @@
+export {default as BottomDetailButton} from './bottomDetailButton';
+export {default as BottomOptionDetailButton} from './bottomOptionDetailButton';
