+++ src/screens/screening/detail/components/detailPositivePlus/index.tsx
